+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### 🚀 New features
+
+- Enabled `TerminalLogger` ([#138](https://github.com/candoumbe/Pipelines/issues/138))
 
 ## [0.9.0] / 2024-01-25
 ### 🚀 New features