+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.10.0] / 2024-07-11
 ### 🚀 New features
 
 - Enabled `TerminalLogger` ([#138](https://github.com/candoumbe/Pipelines/issues/138))
@@ -204,9 +206,8 @@ So now `{MutationTestDirectory}/[{framework}]` is now changed to `{MutationTestD
 ## [0.1.0] / 2022-10-23
 - Initial release
 
-[`EnhancedNukeBuild`]:./src/Candoumbe.Pipelines/Components/EnhancedNukeBuild.cs
-
-[Unreleased]: https://github.com/candoumbe/Pipelines/compare/0.9.0...HEAD
+[Unreleased]: https://github.com/candoumbe/Pipelines/compare/0.10.0...HEAD
+[0.10.0]: https://github.com/candoumbe/Pipelines/compare/0.9.0...0.10.0
 [0.9.0]: https://github.com/candoumbe/Pipelines/compare/0.8.0...0.9.0
 [0.8.0]: https://github.com/candoumbe/Pipelines/compare/0.7.0...0.8.0
 [0.7.0]: https://github.com/candoumbe/Pipelines/compare/0.6.1...0.7.0